services/rules_api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
PyYAML==6.0.2
